Backend/requirements.txt

# =============================================================================
# FIXED REQUIREMENTS.TXT
# =============================================================================
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
twilio==8.10.0
sendgrid==6.10.0
stripe==7.8.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
openai==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1
reportlab==4.0.7
jinja2==3.1.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
pip install pydantic[email]
pip install aiosmtplib emails jinja2
# Add these packages to your backend/requirements.txt

# Integration-specific packages
stripe>=7.0.0
twilio>=8.10.0
sendgrid>=6.10.0
cryptography>=41.0.0
httpx>=0.25.0

# OAuth and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Additional utility packages for integrations
python-multipart>=0.0.6
jinja2>=3.1.2

# Background task processing (optional but recommended)
celery>=5.3.0
redis>=5.0.0

# HTTP client with better async support
aiohttp>=3.9.0


aiofiles==23.2.1  # ✅ NEW - For async file operations
mimetypes==1.0.0  # ✅ NEW - For MIME type detection (built-in but explicit)

google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httpx==0.25.2